requirements.txt
Flask>=1.0.2
GitPython>=2.1.11
gunicorn>=19.9.0
